requirements.txt
PyPDF2
unstructured[pdf]
langdetect
sentence-transformers
google-generativeai
python-dotenv
numpy
langchain
langchain-community
langchain_google_genai
tiktoken
chromadb
pytest
boto3
pillow
detectron2
layoutparser[layoutmodels,tesseract,ocr]
pdfplumber
tabula-py
PyMuPDF